requirements.txt
anyjson==0.3.3
Django==1.8
django-appconf==1.0.1
django-bower==5.0.4
django-compressor==1.5
django-filter==0.10.0
django-rest-framework-mongoengine==2.0.3
django-websocket-redis==0.4.4
djangorestframework==3.1.3
-e git://github.com/dschoeni/drf-mongo-filters.git@268c1735ee761121493a2008dc293bd3819b41cf#egg=drf_mongo_filters-master
gevent==1.0.1
greenlet==0.4.7
mongoengine==0.10.0
pymongo==3.0.1
redis==2.10.3
six==1.9.0
